pyproject.toml
[tool.poetry]
name = "marker-pdf"
version = "0.1.1"
description = "Convert PDF to markdown with high speed and accuracy."
authors = ["Vik Paruchuri <github@vikas.sh>"]
readme = "README.md"
license = "GPL-3.0-or-later"
repository = "https://github.com/VikParuchuri/marker"
keywords = ["pdf", "markdown", "ocr", "nlp"]
packages = [
    {include = "marker"}
]
include = [
    "convert.py",
    "convert_single.py",
    "chunk_convert.sh",
    "benchmark.py",
    "chunk_convert.py",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
scikit-learn = "^1.3.2"
Pillow = "^10.1.0"
pytesseract = "^0.3.10"
PyMuPDF = "^1.23.5"
pymupdf-fonts = "^1.0.5"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
nougat-ocr = "^0.1.17"
transformers = "^4.34.1"
numpy = "^1.26.1"
python-dotenv = "^1.0.0"
torch = "^2.1.1"
ray = "^2.7.1"
tqdm = "^4.66.1"
tabulate = "^0.9.0"
thefuzz = "^0.20.0"
python-magic = "^0.4.27"
pyspellchecker = "^0.7.2"
ftfy = "^6.1.1"
nltk = "^3.8.1"
ocrmypdf = "^15.4.0"
bitsandbytes = "^0.41.2.post2"
grpcio = "^1.60.0"
azure-storage-blob = "^12.19.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.scripts]
marker = "convert:main"
marker_single = "convert_single:main"
marker_benchmark = "benchmark:main"
marker_chunk_convert = "chunk_convert:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"